Create insertionSort.cpp
The Code Successfully implements the Insertion Sort with a Time Complexity of O(n^2) and space complexity of O(1). The Insertion Sort is considered to be an inplace(Requires no extra Memory) and stable( maintains the relative order of the elements). 
Best Case: The algorithm runs with a time Complexity of O(n) when the input array is already sorted.