Create selectionSort.cpp
The code successfully implements the Selection Sort in cpp language. The Selection Sort has a Time Complexity of O(n^2) in all the cases and Space Complexity of O(1). The selesction Sort is considered to be an inplace Sorting Algorithm as it does not requires any additional space but since it does not preserves the relative order of the elements of the input array, it is not a stable Algorithm.