Create Optimized-bubbleSort.cpp
The code optimizes the traditional Bubble Sort by declaring a variable swapped outside of the inner loop. The swapped variable checks if the swapping of the elements has occured in inside loop. The Algorithm gives a Time Complexity of O(n) in Best Case Scenarios(when the input array is already Sorted).