Create Iterative-Binary-Search.cpp
The code successfully implements the Binary Search Algorithm in cpp Language. The Time Complexity of the Algorithm is O(log n) in worst case scenarios but it offers a Time Complexity of O(1) in Best Cases (when the element is found at middle index) .
No Extra Space is required by the Algorithm therefore Binary Search Algorithm has a space Complexity of O(1).